ModuleTypeParser
