+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/ModuleTypeParser.java
@@ -7,18 +7,24 @@ import java.util.Map;
 import com.sap.cloud.lm.sl.common.ParsingException;
 import com.sap.cloud.lm.sl.mta.model.v2.ModuleType;
 import com.sap.cloud.lm.sl.mta.model.v2.ModuleType.Builder;
+import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class ModuleTypeParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ModuleTypeParser {
+public class ModuleTypeParser extends ModelParser<ModuleType> {
 
+    protected static final String PROCESSED_OBJECT_NAME = "MTA module type";
+
+    public static final String NAME = "name";
+    public static final String DEPLOYER = "deployer";
+    public static final String PROPERTIES = "properties";
     public static final String PARAMETERS = "parameters";
 
     public ModuleTypeParser(Map<String, Object> source) {
-        super(MODULE_TYPE, source);
+        this(MODULE_TYPE, source);
     }
 
     protected ModuleTypeParser(MapElement schema, Map<String, Object> source) {
-        super(schema, source);
+        super(PROCESSED_OBJECT_NAME, schema, source);
     }
 
     @Override
@@ -29,6 +35,18 @@ public class ModuleTypeParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ModuleT
         builder.setProperties(getProperties());
         builder.setParameters(getParameters());
         return builder.build();
+    }
+
+    protected String getName() {
+        return getStringElement(NAME);
+    }
+
+    protected String getDeployer() {
+        return getStringElement(DEPLOYER);
+    }
+
+    protected Map<String, Object> getProperties() {
+        return getMapElement(PROPERTIES);
     }
 
     protected Map<String, Object> getParameters() {
